+++ packages/metrics/examples/decorator/constructor-options.ts
@@ -1,12 +1,12 @@
 import * as dummyEvent from '../../../../tests/resources/events/custom/hello-world.json';
 import { context as dummyContext } from '../../../../tests/resources/contexts/hello-world';
 import { Metrics, MetricUnits } from '../../src';
-import { LambdaInterface } from './../utils/lambda';
+import { LambdaInterface } from '../utils/lambda';
 import { Callback, Context } from 'aws-lambda/handler';
 
 const metrics = new Metrics({
   namespace: 'hello-world-constructor',
-  service: 'hello-world-service-constructor'
+  serviceName: 'hello-world-service-constructor'
 });
 
 class Lambda implements LambdaInterface {