+++ packages/tracing/src/Tracer.ts
@@ -21,7 +21,7 @@ import { Segment, Subsegment } from 'aws-xray-sdk-core';
  * 
  * For more usage examples, see [our documentation](https://awslabs.github.io/aws-lambda-powertools-typescript/latest/core/tracer/).
  * 
- * ### Functions usage with middlewares
+ * ### Functions usage with middleware
  * 
  * If you use function-based Lambda handlers you can use the [captureLambdaHandler()](./_aws_lambda_powertools_tracer.Tracer.html) middy middleware to automatically:
  * * handle the subsegment lifecycle 
@@ -257,7 +257,15 @@ class Tracer implements TracerInterface {
   public captureAWSClient<T>(service: T): T {
     if (!this.isTracingEnabled()) return service;
 
-    return this.provider.captureAWSClient(service);
+    try {
+      return this.provider.captureAWSClient(service);
+    } catch (error) {
+      try {
+        return this.provider.captureAWSClient((service as unknown as T & { service: T }).service);
+      } catch {
+        throw error;
+      }
+    }
   }
 
   /**